# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Hoteles Decameron",
        "description": "API para gestionar hoteles, tipos de habitación y acomodaciones.",
        "version": "1.0"
    },
    "paths": {
        "/api/accommodations": {
            "get": {
                "tags": [
                    "Acomodaciones"
                ],
                "summary": "Obtener todas las acomodaciones",
                "description": "Retorna una lista de todas las acomodaciones con sus relaciones (hotel y tipo de habitación).",
                "operationId": "b334c4852bfbfbca991802e1e9ccce4c",
                "responses": {
                    "200": {
                        "description": "Lista de acomodaciones obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Accommodation"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron acomodaciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Acomodaciones"
                ],
                "summary": "Crear una nueva acomodación",
                "description": "Crea una nueva acomodación asociada a un hotel y tipo de habitación.",
                "operationId": "e14fc50aa150f67951581f322d008ab3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Sencilla"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Acomodación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Accommodation"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/accommodations/{id}": {
            "get": {
                "tags": [
                    "Acomodaciones"
                ],
                "summary": "Mostrar una acomodación específica",
                "description": "Retorna los detalles de una acomodación específica por su ID.",
                "operationId": "55273ef755378658b3a42304c8dc64db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la acomodación",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la acomodación obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Accommodation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Acomodación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Acomodaciones"
                ],
                "summary": "Actualizar una acomodación",
                "description": "Actualiza los datos de una acomodación existente.",
                "operationId": "f8385454ae8a5eb33ca72a4dc61bcb9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la acomodación",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Doble"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Acomodación actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Accommodation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Acomodación no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Acomodaciones"
                ],
                "summary": "Eliminar una acomodación",
                "description": "Elimina una acomodación existente por su ID.",
                "operationId": "bb178718927401cff2028d8e0e3dd8e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la acomodación",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Acomodación eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Acomodación eliminada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Acomodación no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Accommodation": {
                "title": "Accommodation",
                "description": "Modelo que representa una acomodación en un hotel.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "hotel_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "room_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "Sencilla"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}